requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
httpx==0.27.2
gspread==6.1.4
google-auth==2.35.0
python-dotenv==1.0.1
pyyaml==6.0.2
